Wrong bit depth

+++ strax/dtypes.py
@@ -161,7 +161,7 @@ def hitlet_with_data_dtype(n_samples):
     additional_fields = [(('Hitlet data in PE/sample with ZLE (only the first length samples are filled)', 'data'),
                            np.float32, n_samples),
                          (('Fragment number in the pulse',
-                           'record_i'), np.int16),
+                           'record_i'), np.int32),
                          ]
 
     return dtype + additional_fields
